Scalamock for ZIO Test
We used ZIO Test and ZIO Mock extensively in our company but wanted to transition away from ZIO Mock. Scalamock was selected as an alternative for the following reasons:
- It is a mature and widely-used library with high-quality error reporting.
- It has been ported to Scala 3.

So, we implemented an integration between the classic Scalamock API and ZIO Test.

This integration was chosen specifically for several reasons:
- It makes migration from ZIO Mock easier compared to adopting the experimental Scalamock API.
- Our developers are already familiar with this API.

Additionally, a few recurring issues with classic API have been addressed:
- In the classic Scalamock API, it’s too easy to create shared mocks across multiple tests, which can lead to race conditions. In this integration, we handle the creation of mocks through ZLayer, eliminating this issue.
- The classic API does not provide a way to verify whether an effect was actually called. This integration introduces that capability.

I believe these considerations are not only relevant to our team, but could also be helpful for other users. Specifically, those looking to transition from ZIO Mock to Scalamock or use the classic Scalamock API with ZIO Test.